fix 실내 지도, 실외 지도 분리 및 전환 버튼 추가
